refactor: Statementクラスを抽出
